Remove PostProcessing dependency from Core
