+++ com.unity.render-pipelines.core/CoreRP/Editor/com.unity.render-pipelines.core.Editor.asmdef
@@ -1,9 +1,7 @@
 {
     "name": "com.unity.render-pipelines.core.Editor",
     "references": [
-        "com.unity.render-pipelines.core.Runtime",
-        "com.unity.postprocessing.Runtime",
-        "com.unity.postprocessing.Editor"
+        "com.unity.render-pipelines.core.Runtime"
     ],
     "optionalUnityReferences": [],
     "includePlatforms": [
