+++ com.unity.render-pipelines.core/CoreRP/com.unity.render-pipelines.core.Runtime.asmdef
@@ -1,8 +1,5 @@
 {
     "name": "com.unity.render-pipelines.core.Runtime",
-    "references": [
-        "com.unity.postprocessing.Runtime"
-    ],
     "optionalUnityReferences": [],
     "includePlatforms": [],
     "excludePlatforms": [],
