+++ com.unity.render-pipelines.core/package.json
@@ -3,8 +3,5 @@
     "description": "Core library for Unity render pipelines.",
     "version": "3.0.0-preview",
     "unity": "2018.2",
-    "displayName": "Render Pipeline Core Library",
-    "dependencies": {
-        "com.unity.postprocessing": "2.0.7-preview"
-    }
+    "displayName": "Render Pipeline Core Library"
 }
